Move light client and relayer tasks into their own modules
