+++ relayer-cli/src/tasks.rs
@@ -1,5 +1,5 @@
 //! Tasks which make up the relayer application
 
 pub mod event_listener;
-// pub mod light_client;
-// pub mod relayer;
+pub mod light_client;
+pub mod relayer;
